Rename attribute, struct and files
