+++ Cargo.toml
@@ -15,3 +15,5 @@ chrono = "0.4"
 petgraph = "0.6"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
+csv = "1"
+rand = "0"
